[AIC-py][editor] split server into 2 files
This is getting kind of unweildy. Move most of the hleper functions into server_utils,
leaving the core server stuff (mainly endpoints).
